refactor: `CartList`의 관심사 분리
